% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_equi_X.R
\name{create_equi_X}
\alias{create_equi_X}
\title{Creating knockoff matrix}
\usage{
create_equi_X(X)
}
\arguments{
\item{X}{the data as a matrix requiring n>=2p}
}
\value{
An object of class "knockoff.variables". This is a list containing at least the following components:\code{n}

X n-by-p matrix of original variables(standardized).

Xk	n-by-p matrix of knockoff variables.

s_j 2*the minimum eigenvalue of 'Sigma'
}
\description{
Creating knockoff matrix using Equi-correlated knockoffs
}
\examples{
\dontrun{
n2 <- 200;p2 <- 100
X2 = matrix(rnorm(n2*p2),n2)
X2q <- create.equi.X(X2)}
}
\references{
Barber and Candes, Controlling the false discovery rate via knockoffs. Ann. Statist. 43 (2015), no. 5, 2055–2085
}
