% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jack.R
\name{jack}
\alias{jack}
\title{jackknife method}
\usage{
jack(data, fun = NULL)
}
\arguments{
\item{data}{the data given as a vector}

\item{fun}{function to be bootstrapped}
}
\value{
the standard error to be estimated
}
\description{
Compute the jackknife estimate of standard error using R
}
\examples{
\dontrun{
data <- 20 * rnorm(500,2,3)
jack(data = data, fun = mean)
}
}
