% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rw_Metropolis}
\alias{rw_Metropolis}
\alias{dLaplace}
\title{A Metropolis sampler using Rcpp}
\usage{
dLaplace(x)
}
\arguments{
\item{N}{the number of samples}

\item{x0}{the number of between-sample random number}

\item{sigma}{variance}
}
\value{
a List with a random sample of size \code{n}
}
\description{
A Metropolis sampler using Rcpp
}
